DHS-152 Reorg repository for package scanning
