+++ src/main/java/gov/dhs/nppd/humanreview/service/UserApiService.java
@@ -7,7 +7,7 @@ import javax.validation.Valid;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.openapitools.model.AuthCredentials;
-import org.openapitools.repository.AuthCredentialsRepository;
+import org.openapitools.repository.auth.AuthCredentialsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
