+++ src/main/java/gov/dhs/nppd/humanreview/util/DB1_DataSource.java
@@ -17,7 +17,7 @@ import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
 import org.springframework.transaction.PlatformTransactionManager;
 
 @Configuration
-@EnableJpaRepositories(basePackages = {"org.openapitools.repository"},
+@EnableJpaRepositories(basePackages = {"org.openapitools.repository.auth"},
         entityManagerFactoryRef = "db1EntityManager",
         transactionManagerRef = "db1TransactionManager")
 public class DB1_DataSource {
@@ -34,7 +34,7 @@ public class DB1_DataSource {
         em.setJpaVendorAdapter(vendorAdapter);
         HashMap<String, Object> properties = new HashMap<>();
         properties.put("hibernate.dialect",
-                env.getProperty("hibernate.dialect"));
+                env.getProperty("spring.jooq.sql-dialect"));
         properties.put("hibernate.show-sql",
                 env.getProperty("jdbc.show-sql"));
         em.setJpaPropertyMap(properties);
