+++ src/test/java/gov/dhs/nppd/humanreview/service/HumanreviewApiServiceTest.java
@@ -16,8 +16,8 @@ import org.mockito.Mockito;
 import org.openapitools.model.HumanReviewItem;
 import org.openapitools.model.JsonData;
 import org.openapitools.model.ListOfHumanReviewItems;
-import org.openapitools.repository.HumanreviewRepository;
-import org.openapitools.repository.JsonDataRepository;
+import org.openapitools.repository.humanreview.HumanreviewRepository;
+import org.openapitools.repository.humanreview.JsonDataRepository;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
