+++ src/test/java/gov/dhs/nppd/humanreview/service/UserApiServiceTest.java
@@ -8,7 +8,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.openapitools.model.AuthCredentials;
-import org.openapitools.repository.AuthCredentialsRepository;
+import org.openapitools.repository.auth.AuthCredentialsRepository;
 import org.springframework.http.ResponseEntity;
 
 import gov.dhs.nppd.humanreview.controller.UserApiController;
